pyproject.toml
[project]
name = "omnibot"
version = "0.1.0"
description = "A general all-purpose discord bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
